SSH bug related to the provisioning of controller identifiers fixed.
Due to ill-tree conditions in non-loop-free topologies, DHCP messages are duplicated.
This leads to opening of multiple SSH connections between a controller and the switch
that broadcasts its DHCP request messages into the ill tree. DHCP server handles each
duplicate separately. That is, for each request message it offers a new IP address to
the switch. The switch can accept a new IP address while already establishing an SSH connection
with the old one. In this case, the existing SSH/OF connection would break and the switch
would stuck in this state forever. However, if dhclient is manually triggered on the problematic
switch, the bootstrapping procedure continues to work as usual. Thus, we try to handle each
SSH attempt in the same manner by trying to deliver controller identifiers as long as it is
necessary to succeed. The SSH thread that succeeds will stop other threads that try the same
only with other IP addresses. The better solution would be to configure/implement the DHCP
server/client to not to react on duplicates. For this, a more insightful look into the DHCP
implementation is required.